Nicer layout for CDN uploading in admin + files which error will now hang around for review

+++ asset/assets/js/nails.admin.cdn.upload.js
@@ -27,7 +27,7 @@ NAILS_Admin_CDN_Upload = function()
 	{
 		var _this = this;
 		Dropzone.autoDiscover = false;
-		this._dropzone = new Dropzone("div#dropzone",
+		this._dropzone = new Dropzone('div#dropzone',
 		{
 			url: window.SITE_URL + 'api/cdn/object/create',
 			autoProcessQueue: true,
@@ -40,12 +40,19 @@ NAILS_Admin_CDN_Upload = function()
 		});
 
 		//	Bind listeners
-		this._dropzone.on("complete", function(file) {
-			_this._dropzone.removeFile(file);
+		this._dropzone.on('complete', function(file) {
+
+			if (file.status === 'success') {
+				_this._dropzone.removeFile(file);
+			}
 		});
-		this._dropzone.on("queuecomplete", function() {
+		this._dropzone.on('queuecomplete', function() {
+
 			$('#alert-complete').slideDown();
-			setTimeout( function(){$('#alert-complete').slideUp();}, 7500 );
+			setTimeout(function() {
+
+				$('#alert-complete').slideUp();
+			}, 7500);
 		});
 
 		$('#action-upload').on('click', function()
